Move timers into ubrn_testing.
