+++ crates/ubrn_testing/src/lib.rs
@@ -4,3 +4,4 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/
  */
 pub mod thread;
+pub mod timer;
