+++ crates/ubrn_testing/src/timer/mod.rs
@@ -3,13 +3,21 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/
  */
+use std::{future::Future, time::Duration};
+
+pub struct TimerFuture;
+pub trait TimerService {
+    type Future: Future;
+    fn sleep(duration: Duration) -> Self::Future;
+}
 
 #[cfg(not(target_arch = "wasm32"))]
 mod native;
 #[cfg(not(target_arch = "wasm32"))]
-pub(crate) use native::*;
+pub use native::*;
 
 #[cfg(target_arch = "wasm32")]
 mod wasm;
+#[allow(unused)]
 #[cfg(target_arch = "wasm32")]
 pub(crate) use wasm::*;
