+++ fixtures/async-callbacks/Cargo.toml
@@ -10,21 +10,17 @@ publish = false
 name = "async_callbacks"
 crate-type = ["lib", "cdylib"]
 
-[[bin]]
-name = "uniffi-fixtures-async-callbacks"
-path = "src/bin.rs"
-
 [dependencies]
-# Add the "scaffolding-ffi-buffer-fns" feature to make sure things can build correctly
-uniffi = { workspace = true, features = ["tokio", "cli", "scaffolding-ffi-buffer-fns"] }
+uniffi = { workspace = true, features = ["tokio", "cli"] }
 async-trait = "0.1"
 futures = "0.3"
 thiserror = "1.0"
 tokio = { version = "1.24.1", features = ["time", "sync"] }
 once_cell = "1.18.0"
+ubrn_testing = { path = "../../crates/ubrn_testing" }
 
 [build-dependencies]
-uniffi = { workspace = true, features = ["build", "scaffolding-ffi-buffer-fns"] }
+uniffi = { workspace = true, features = ["build"] }
 
 [dev-dependencies]
 uniffi = { workspace = true, features = ["bindgen-tests"] }
