--- fixtures/async-callbacks/src/async-callbacks.udl
@@ -1,21 +1,0 @@
-namespace futures {
-    [Async]
-    boolean always_ready();
-};
-
-[Trait]
-interface SayAfterUdlTrait {
-    [Async]
-    string say_after(u16 ms, string who);
-};
-
-interface UdlMegaphone {
-    [Async]
-    constructor();
-
-    [Async, Name="secondary"]
-    constructor();
-
-    [Async]
-    string say_after(u16 ms, string who);
-};
