--- fixtures/async-callbacks/src/bin.rs
@@ -1,8 +1,0 @@
-/*
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/
- */
-fn main() {
-    uniffi::uniffi_bindgen_main()
-}
