+++ fixtures/async-calls/Cargo.toml
@@ -20,10 +20,8 @@ async-trait = "0.1"
 futures = "0.3"
 thiserror = "1.0"
 once_cell = "1.18.0"
+ubrn_testing = { path = "../../crates/ubrn_testing" }
 uniffi = { workspace = true, features = ["tokio", "cli", "scaffolding-ffi-buffer-fns"] }
-
-[target.'cfg(target_arch = "wasm32")'.dependencies]
-gloo-timers = { version = "0.3.0", features = ["futures"] }
 
 [target.'cfg(not(target_arch = "wasm32"))'.dependencies]
 tokio = { version = "1.24.1", features = ["time", "sync"] }
